Extract FakeUsersRepository and TestUsers to :data:users:core-test module
I plan to use FakeUsersRepository and TestUsers in :ui:userdetails - extracting to separate module
will help me reuse it there.
