+++ data/users/core-test/build.gradle.kts
@@ -1,0 +1,6 @@
+plugins {
+    id("kotlin")
+}
+dependencies {
+    api(project(":data:users:core"))
+}
